backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
requests
python-dotenv
python-multipart
transformers
accelerate
torch
